% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{blogosphere}
\alias{blogosphere}
\title{blogosphere}
\usage{
data(blogosphere)
}
\description{
blogosphere network
}
\details{
blogosphere network, in \code{igraph}-format.

The network consists of 1490 blogsites connected by 19090 directed edges
(one blogsite linking to another).

The vertex attributes are
\itemize{
\item{id}{id of the blogsite}
\item{label}{url of the blog}
\item{value}(numeric, political preference of the blog: 0 for liberal/democrat, 1 for conservative/republican)
\item{source}{character, weblog directories}
\item{party}{character, "liberal" or "conservative"}
\item{color}{character, "blue" for liberal, "red" for conservative}
}

The edge attributes are
\itemize{
\item{within}{logical, whether the edge connects two blogs with the same political point-of-view}
\item{within_republican}{logical, whether the edge connects two republican blogsites}
\item{within_democrat}(logical, whether the edge connects two democratic blogsites)
}

Note: there are loops, since some blogs refer to themselves.
}
\references{
,
Lada Adamic & Natalie Glance, The political blogosphere and the 2004 U.S.
election: Divided they blog, 2005, LinkeKDD '05: Proceedings of the 3rd
international workshop op Link discovery.
}
\keyword{datasets}
