% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DSstudents}
\alias{DSstudents}
\title{Data Science Students}
\format{
An igraph object with 81 nodes and 148 edges. Directed, Weighted.
}
\source{
Data Collected during the SNA4DS course 2021 and modified for anonymization
}
\usage{
data(DSstudents)
}
\description{
DSstudents
}
\details{
DSstudents, in \code{igraph}-format.
}
\references{
Zucca, C., 2021, Data Science Students -SNA4DS 2021
}
\keyword{datasets}
