% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{elegans}
\alias{elegans}
\title{elegans network data}
\format{
A data frame with 297 vertices and 2344 edges. Directed, Weighted.
}
\source{
\url{http://networkrepository.com/celegansneural.php}
}
\usage{
data(elegans)
}
\description{
elegans network
}
\details{
elegans network, in \code{igraph}-format.
Neural network of the nematode Caenorhabditis Elegans.
So, nodes are neurons, links are outgoing or incoming synapses.
}
\references{
White JG, Southgate E, Thomson JN and S Brenner, 1986,
The structure of the nervous system of the nematode Caenorhabditis elegans,
Philos Trans R Soc Lond B Biol Sci., pages 1-340.
}
\keyword{datasets}
