% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fifa2006}
\alias{fifa2006}
\title{FIFA 2006 committee membership network}
\usage{
data(fifa2006, package = "SNA4DSData")
}
\description{
FIFA committee membership network in 2006
}
\details{
Network (in \code{network} format) of co-membership in committees
in FIFA.The vertices are
members of the FIFA committees and the edges represent the whether
the members of the dyad are a member of a least one joint committee.

This is the network for the year 2006.
}
\keyword{fifa2006}
