% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fifa2015}
\alias{fifa2015}
\title{FIFA 2015 committee membership network}
\usage{
data(fifa2015, package = "SNA4DSData")
}
\description{
FIFA committee membership network in 2015
}
\details{
Network (in \code{network} format) of co-membership in committees
in FIFA.The vertices are
members of the FIFA committees and the edges represent the whether
the members of the dyad are a member of a least one joint committee.

This is the network for the year 2015.
}
\keyword{fifa2015}
