% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ia_enron}
\alias{ia_enron}
\title{ia_enron network data}
\format{
143 vertices and 622 edges. Directed, Weighted.
}
\usage{
data(ia_enron)
}
\description{
ia_enron network
}
\details{
ia_enron network, in \code{igraph}-format.
}
\keyword{datasets}
