% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{imdb_costars}
\alias{imdb_costars}
\title{imdb costars}
\format{
A data frame with 252999 vertices and 1015187 edges. Undirected, Unweighted.
}
\source{
\url{https://www.imdb.com/}
}
\usage{
data(imdb_costars)
}
\description{
imdb costars network
}
\details{
imdb costars network, in \code{igraph}-format.
An edge is this network is defined as the actors being part of the same movie.
The edge weight attribute reflects the number of collaborations.

I added the actual names of the actors by scraping the imdb website accordingly.
}
\keyword{datasets}
