% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{louis}
\alias{louis}
\title{Louisiana political preferences}
\format{
A data frame with 64 vertices and 302 edges. Undirected, unweighted.
}
\usage{
data(louis, package = "SNA4DSData")
}
\description{
Louisiana political preferences
}
\details{
These data come from a study on the determinants of presidential
voting behavior in the parishes of Louisiana (1980).
Here, the dependent variable of interest is the percentage of the voting
electorate who support particular Democratic presidential candidates.

Several dimensions that are important characteristics of Louisiana parishes
include percentage black, percentage Catholic, the
percentage black political equality, and the percentage urban.
Each of these variables, in the aggregate, can be seen as predictors of
electoral turnout and partisan electoral behavior in presidential elections.

The network represents contiguity between the counties.
}
\keyword{datasets}
