% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{northwestern}
\alias{northwestern}
\title{Northwestern University Facebook network}
\format{
A data frame with 10567 vertices and 488336 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/socfb-Northwestern25.php}
}
\usage{
data(northwestern, package = "SNA4DSData")
}
\description{
Northwestern University Facebook network
}
\details{
Northwestern University Facebook network, in \code{igraph}-format.

A social friendship network extracted from Facebook consisting of people
(nodes) with edges representing "friendship" ties.

There are 10567 Facebook users with 488336 friendship edges between them.
}
\references{
Traud, Amanda L and Mucha, Peter J and Porter, Mason A,
Social structure of Facebook networks, 2012,
Phys. A, August, 16, 391: 4165-4180.

Traud, Amanda L and Kelsic, Eric D and Mucha, Peter J and Porter, Mason A, 2011,
Comparing Community Structure to Characteristics in Online Collegiate Social Networks,
SIAM Rev., 53 (3): 526--543
}
\keyword{datasets}
