% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{routers}
\alias{routers}
\title{routers network}
\format{
A data frame with 190914 vertices and 607610 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/tech-RL-caida.php}
}
\usage{
data(routers, package = "SNA4DSData")
}
\description{
routers network
}
\details{
routers network, in \code{igraph}-format. The network is known as RL-CAIDA.
This represents the structure of the internet (well, a snapshot of a
small part of it).

190914 routers linked by 607610 undirected edges.
}
\references{
Ryan A. Rossi, Sonia Fahmy, and Nilothpal Talukder, 2013,
A Multi-Level Approach for Evaluating Internet Topology Generators,
IFIP Networking, pages 1-9.

Spring, Neil and Mahajan, Ratul and Wetherall, David, 2002,
Measuring ISP topologies with Rocketfuel,
ACM SIGCOMM Computer Communication Review, 32 (4): 133--145.
}
\keyword{datasets}
