% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4DSData-package.r
\docType{package}
\name{SNA4DSData}
\alias{SNA4DSData}
\alias{SNA4DSData-package}
\title{SNA4DSData}
\description{
Package contaning datasets for the SNA4DS course
}
\details{
Datasets, datasets, and datasets, useful for the SNA4DS course.
}
\author{
\strong{Maintainer}: Roger Leenders \email{SNA4DS@jads.nl}

Authors:
\itemize{
  \item Claudia Zucca \email{SNA4DS@jads.nl}
}

}
\keyword{internal}
\keyword{package}
