% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_air}
\alias{us_air}
\title{us air network}
\format{
A data frame with 546 vertices and 2781 edges. Undirected, unweighted.
}
\usage{
data(us_air, package = "SNA4DSData")
}
\description{
us air network
}
\details{
us air network, in \code{igraph}-format.

The data were generated from the Feb 2, 2017 data release \url{https://github.com/jpatokal/openflights/tree/5d623a6969a1adee7961cf1c9a8a212c4a784713}.

This is a subset of the world air network dataset.
}
\keyword{datasets}
