% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{yeast network data}
\format{
A data frame with 1870 vertices and 4480 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/bio-yeast-protein-inter.php}
}
\usage{
data(yeast, package = "SNA4DSData")
}
\description{
Yeast network
}
\details{
Yeast protein network, in \code{igraph}-format, describing interacting proteins.

There are 1870 proteins with 4480 interactions.
}
\references{
Jeong, H. and Mason, S.P. and Barabasi, A.L. and Oltvai, Z.N., 2001,
Lethality and centrality in protein networks, arXiv preprint cond-mat0105306.
}
\keyword{datasets}
